rename specs to match reflect new "onboarding" directory
