+++ tutor/specs/components/onboarding/expired-preview-warning.spec.jsx
@@ -1,10 +1,9 @@
 import { Wrapper, SnapShot } from '../helpers/component-testing';
-import SecondSessionWarning from '../../../src/components/course-preview/second-session-warning';
+import ExpiredPreviewWarning from '../../../src/components/onboarding/expired-preview-warning';
 import CoursePreviewUX from '../../../src/models/course/preview-ux';
 import EnzymeContext from '../helpers/enzyme-context';
-import { extend } from 'lodash';
 
-describe('Second Session Warning', () => {
+describe('Expired Preview Warning', () => {
 
   let ux;
   beforeEach(() => {
@@ -14,12 +13,12 @@ describe('Second Session Warning', () => {
 
   it('renders and matches snapshot', () => {
     expect(SnapShot.create(
-      <Wrapper _wrapped_component={SecondSessionWarning} ux={ux} />).toJSON()
+      <Wrapper _wrapped_component={ExpiredPreviewWarning} ux={ux} />).toJSON()
     ).toMatchSnapshot();
   });
 
   it('dislays got it and dismisses on continue', () => {
-    const warning = shallow(<SecondSessionWarning ux={ux} />);
+    const warning = shallow(<ExpiredPreviewWarning ux={ux} />);
     warning.find('Button[bsStyle="default"]').simulate('click');
     expect(warning.find('Body').render().text()).toContain('No problem');
     warning.find('Button[bsStyle="primary"]').simulate('click');
@@ -28,7 +27,7 @@ describe('Second Session Warning', () => {
 
   it('navigates on add', () => {
     const context =  EnzymeContext.build();
-    const warning = shallow(<SecondSessionWarning ux={ux} />, context);
+    const warning = shallow(<ExpiredPreviewWarning ux={ux} />, context);
     warning.find('Button[bsStyle="primary"]').simulate('click');
     expect(context.context.router.transitionTo).to.have.been.calledWith('/dashboard');
   });
