+++ tutor/specs/components/onboarding/second-session-warning.spec.jsx
@@ -1,9 +1,10 @@
 import { Wrapper, SnapShot } from '../helpers/component-testing';
-import ExpiredPreviewWarning from '../../../src/components/course-preview/expired-preview-warning';
+import SecondSessionWarning from '../../../src/components/onboarding/second-session-warning';
 import CoursePreviewUX from '../../../src/models/course/preview-ux';
 import EnzymeContext from '../helpers/enzyme-context';
+import { extend } from 'lodash';
 
-describe('Expired Preview Warning', () => {
+describe('Second Session Warning', () => {
 
   let ux;
   beforeEach(() => {
@@ -13,12 +14,12 @@ describe('Expired Preview Warning', () => {
 
   it('renders and matches snapshot', () => {
     expect(SnapShot.create(
-      <Wrapper _wrapped_component={ExpiredPreviewWarning} ux={ux} />).toJSON()
+      <Wrapper _wrapped_component={SecondSessionWarning} ux={ux} />).toJSON()
     ).toMatchSnapshot();
   });
 
   it('dislays got it and dismisses on continue', () => {
-    const warning = shallow(<ExpiredPreviewWarning ux={ux} />);
+    const warning = shallow(<SecondSessionWarning ux={ux} />);
     warning.find('Button[bsStyle="default"]').simulate('click');
     expect(warning.find('Body').render().text()).toContain('No problem');
     warning.find('Button[bsStyle="primary"]').simulate('click');
@@ -27,7 +28,7 @@ describe('Expired Preview Warning', () => {
 
   it('navigates on add', () => {
     const context =  EnzymeContext.build();
-    const warning = shallow(<ExpiredPreviewWarning ux={ux} />, context);
+    const warning = shallow(<SecondSessionWarning ux={ux} />, context);
     warning.find('Button[bsStyle="primary"]').simulate('click');
     expect(context.context.router.transitionTo).to.have.been.calledWith('/dashboard');
   });
